backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
requests==2.31.0
python-multipart==0.0.6
numpy==1.26.2
scipy==1.11.4
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
earthengine-api==0.1.359
python-dotenv==1.0.0
cors==1.0.1